+++ app/src/main/java/com/example/mukesh/trackyourvehicle/TripCalculate.java
@@ -29,100 +29,100 @@ public class TripCalculate
 
 
    //1.
-    public static float getTotalTripDistance(float distance_OJ,float distance_RJ,float distance_no_load)
+    public static double getTotalTripDistance(double distance_OJ,double distance_RJ,double distance_no_load)
     {
         return distance_OJ + distance_RJ + distance_no_load;
     }
 
     //2
-    public static float getTotalTripLoadCarried(float OJ_Load,float RJ_Load)
+    public static double getTotalTripLoadCarried(double OJ_Load,double RJ_Load)
     {
         return OJ_Load + RJ_Load;
     }
 
     //3
-    public static float getMaintenanceCostPerKmPerYear(float cost_maintenance_year,float total_distance_year)
+    public static double getMaintenanceCostPerKmPerYear(double cost_maintenance_year,double total_distance_year)
     {
         return cost_maintenance_year/ total_distance_year;
     }
 
     //4
-    public static float getRunningCostPerYear(float cost_fuel_year,float cost_maintenance_year)
+    public static double getRunningCostPerYear(double cost_fuel_year,double cost_maintenance_year)
     {
         return cost_fuel_year + cost_maintenance_year;
     }
 
     //5
-    public static float getMaintenanceCostPerYear(float cost_maintenance_km,float total_distance_year)
+    public static double getMaintenanceCostPerYear(double cost_maintenance_km,double total_distance_year)
     {
         return cost_maintenance_km * total_distance_year;
     }
 
     //6
-    public static float getDistancePerYear(float no_trip_year, float trip_distance)
+    public static double getDistancePerYear(double no_trip_year, double trip_distance)
     {
         return no_trip_year * trip_distance;
     }
 
     //7
-    public static float getFuelCostPerYear(float price_fuel,float distance_year,float trip_mileage_km)
+    public static double getFuelCostPerYear(double price_fuel,double distance_year,double trip_mileage_km)
     {
         return price_fuel * (distance_year / trip_mileage_km);
     }
 
     //8
-    public static float getTripMileage(float total_trip_distance,float trip_distance_load,float mileage_load,float distance_without_load,float mileage_without_load)
+    public static double getTripMileage(double total_trip_distance,double trip_distance_load,double mileage_load,double distance_without_load,double mileage_without_load)
     {
         return total_trip_distance / ( (trip_distance_load / mileage_load) + (distance_without_load / mileage_without_load) );
     }
 
     //9
-    public static float getDistancePerMonth(float no_trips_month,float distance_trip)
+    public static double getDistancePerMonth(double no_trips_month,double distance_trip)
     {
         return no_trips_month * distance_trip;
     }
 
     //10
-    public static float getPayLoadInTonsPerYear(float total_load_trip,float trip_year)
+    public static double getPayLoadInTonsPerYear(double total_load_trip,double trip_year)
     {
         return total_load_trip * trip_year;
     }
 
     //11
-    public static float getTotalLoadPerTrip(float onward_load,float return_load)
+    public static double getTotalLoadPerTrip(double onward_load,double return_load)
     {
         return onward_load + return_load;
     }
 
     //12
-    public static float getTotalTonKmPerYear(float onward_load,float onward_distance,float return_load,float return_distance,float no_trips_month,float no_month)
+    public static double getTotalTonKmPerYear(double onward_load,double onward_distance,double return_load,double return_distance,double no_trips_month,double no_month)
     {
         return ((onward_load * onward_distance) + (return_load * return_distance)) * no_trips_month * no_month;
     }
 
 
     //13
-    public static float getTotalFreightEarnedPerYear(float freight_rate_ton_km_OJ,float load_OJ,float distance_load_OJ,float freight_rate_ton_km_RJ,float load_RJ,float distance_load_RJ,float no_trips_year)
+    public static double getTotalFreightEarnedPerYear(double freight_rate_ton_km_OJ,double load_OJ,double distance_load_OJ,double freight_rate_ton_km_RJ,double load_RJ,double distance_load_RJ,double no_trips_year)
     {
         return ((freight_rate_ton_km_OJ * load_OJ * distance_load_OJ) + (freight_rate_ton_km_RJ * load_RJ * distance_load_RJ)) * no_trips_year;
     }
 
 
     //15
-    public static float getInitialCost(float body_cost,float vehicle_price)
+    public static double getInitialCost(double body_cost,double vehicle_price)
     {
         return body_cost + vehicle_price;
     }
 
 
     //16
-    public static float getCrewSalaryYear(float operative_moth,float crew_salary_month)
+    public static double getCrewSalaryYear(double operative_moth,double crew_salary_month)
     {
         return operative_moth * crew_salary_month;
     }
 
     //16
-    public static float getTotalFixedCost(float crew_salary_year,float taxes ,float admin_expenses)
+    public static double getTotalFixedCost(double crew_salary_year,double taxes ,double admin_expenses)
     {
         return crew_salary_year * taxes * admin_expenses;
     }
